Add TextureRegion and use it in Sprite
A TextureRegion stores a texture and offset + size in the texture,
i.e. a region in the texture. Sprites now have a TextureRegion
instead of storing this information directly. Now just need to add
a drawing method for TextureRegions that lets the caller specify
the location and size to draw.